Start rewriting test suite in python
It's just so much better suited to this, wow.

Also I'm doing away with the tar blob in favour of dynamically
generating the git repo each time the test suite runs. Unlike the code
used for the old go tests, all of this is automatically stored in
temporary folders that are automatically cleaned up after the tests are
finished. This means it's safe to run outside of a container too.

Also in general the git repo generator has been made significantly more
resilient against various git configurations. It now doesn't matter if
you sign your commits or tags, or what your author/committer details
are, or even when you run the tests.

Finally, various bash safety stuff has been turned on in some scripts.